Silencing most compiler warnings from GCC 4.1.
The only compiler warning I didn't silence was of this form:
"warning: 'foo' has a base '<unnamed>::bar' whose type uses the anonymous namespace

This appears to be spurious and fixed in later versions of GCC: http://gcc.gnu.org/bugzilla/show_bug.cgi?id=29365

The other warning were fairly innocuous, but useful to hammer out as a first commit I've done here :-)

sc_heal.cpp - needed a newline at the end of the file 
sc_sim.cpp - use fabs instead of abs to avoid some unecessary double <-> int conversions
sc_thread.cpp - need to pass unsigned int instead of double to sleep (if you've overflowed the unsigned int, you'll be sleeping a long time anyway)
sc_sample_data.hpp - this was interesting, was converting to result_type prematurely, and then calculating the sum, min, and max using data_type instead. All the template class instantiations I see are unsigned int and double so it doesn't look like it was possible for this to cause numerical errors, just slower code doing unnecessary conversions. 



git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@14584 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_thread.cpp
@@ -122,7 +122,7 @@ public:
   void join() { pthread_join( t, NULL ); }
 
   static void sleep( timespan_t t )
-  { ::sleep( t.total_seconds() ); }
+    { ::sleep( (unsigned int)t.total_seconds() ); }
 };
 
 #else
